backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
alembic
httpx
python-jose[cryptography]
python-dotenv
